So long, dtslint expect
