--- packages/dtslint/dtslint-expect-only.json
@@ -1,6 +1,0 @@
-{
-    "rulesDirectory": "./dist/rules",
-    "rules": {
-        "expect": true
-    }
-}
