+++ packages/dtslint/dtslint.json
@@ -1,7 +1,6 @@
 {
     "rulesDirectory": "./dist/rules",
     "rules": {
-        "expect": true,
         "npm-naming": true
     }
 }
