+++ packages/dtslint/src/lint.ts
@@ -2,14 +2,13 @@ import { TypeScriptVersion } from "@definitelytyped/typescript-versions";
 import { typeScriptPath, withoutStart } from "@definitelytyped/utils";
 import assert = require("assert");
 import fs from "fs";
-import { join as joinPaths, normalize } from "path";
+import { join as joinPaths, normalize, dirname, resolve as resolvePath } from "path";
 import { Configuration, Linter } from "tslint";
 import { ESLint } from "eslint";
 import * as TsType from "typescript";
 type Configuration = typeof Configuration;
 type IConfigurationFile = Configuration.IConfigurationFile;
 
-import { getProgram, Options as ExpectOptions } from "./rules/expectRule";
 import { readJson } from "./util";
 
 export async function lint(
@@ -34,10 +33,10 @@ export async function lint(
   }
 
   const linter = new Linter({ fix: false, formatter: "stylish" }, lintProgram);
-  const configPath = expectOnly ? joinPaths(__dirname, "..", "dtslint-expect-only.json") : getConfigPath(dirPath);
+  const configPath = getConfigPath(dirPath);
   // TODO: To port expect-rule, eslint's config will also need to include [minVersion, maxVersion]
   //   Also: expect-rule should be renamed to expect-type or check-type or something
-  const config = getLintConfig(configPath, tsconfigPath, minVersion, maxVersion, tsLocal);
+  const config = getLintConfig(configPath);
   const esfiles = [];
 
   for (const file of lintProgram.getSourceFiles()) {
@@ -222,13 +221,7 @@ function getConfigPath(dirPath: string): string {
   return joinPaths(dirPath, "tslint.json");
 }
 
-function getLintConfig(
-  expectedConfigPath: string,
-  tsconfigPath: string,
-  minVersion: TsVersion,
-  maxVersion: TsVersion,
-  tsLocal: string | undefined,
-): IConfigurationFile {
+function getLintConfig(expectedConfigPath: string): IConfigurationFile {
   const configExists = fs.existsSync(expectedConfigPath);
   const configPath = configExists ? expectedConfigPath : joinPaths(__dirname, "..", "dtslint.json");
   // Second param to `findConfiguration` doesn't matter, since config path is provided.
@@ -237,18 +230,6 @@ function getLintConfig(
     throw new Error(`Could not load config at ${configPath}`);
   }
 
-  const expectRule = config.rules.get("expect");
-  if (!expectRule || expectRule.ruleSeverity !== "error") {
-    throw new Error("'expect' rule should be enabled, else compile errors are ignored");
-  }
-  if (expectRule) {
-    const versionsToTest = range(minVersion, maxVersion).map((versionName) => ({
-      versionName,
-      path: typeScriptPath(versionName, tsLocal),
-    }));
-    const expectOptions: ExpectOptions = { tsconfigPath, versionsToTest };
-    expectRule.ruleArguments = [expectOptions];
-  }
   return config;
 }
 
@@ -274,3 +255,50 @@ function range(minVersion: TsVersion, maxVersion: TsVersion): readonly TsVersion
 }
 
 export type TsVersion = TypeScriptVersion | "local";
+
+// TODO(jakebailey): this is also copied into the eslint-plugin; find a way to share it or remove.
+const programCache = new WeakMap<TsType.Program, Map<string, TsType.Program>>();
+/** Maps a tslint Program to one created with the version specified in `options`. */
+export function getProgram(
+  configFile: string,
+  ts: typeof TsType,
+  versionName: string,
+  lintProgram: TsType.Program,
+): TsType.Program {
+  let versionToProgram = programCache.get(lintProgram);
+  if (versionToProgram === undefined) {
+    versionToProgram = new Map<string, TsType.Program>();
+    programCache.set(lintProgram, versionToProgram);
+  }
+
+  let newProgram = versionToProgram.get(versionName);
+  if (newProgram === undefined) {
+    newProgram = createProgram(configFile, ts);
+    versionToProgram.set(versionName, newProgram);
+  }
+  return newProgram;
+}
+
+function createProgram(configFile: string, ts: typeof TsType): TsType.Program {
+  const projectDirectory = dirname(configFile);
+  const { config } = ts.readConfigFile(configFile, ts.sys.readFile);
+  const parseConfigHost: TsType.ParseConfigHost = {
+    fileExists: fs.existsSync,
+    readDirectory: ts.sys.readDirectory,
+    readFile: (file) => fs.readFileSync(file, "utf8"),
+    useCaseSensitiveFileNames: true,
+  };
+  const parsed = ts.parseJsonConfigFileContent(config, parseConfigHost, resolvePath(projectDirectory), {
+    noEmit: true,
+  });
+
+  if (config.compilerOptions?.module === "node16" && parsed.options.module === undefined) {
+    // TypeScript version is too old to handle the "node16" module option,
+    // but we can run tests falling back to commonjs/node.
+    parsed.options.module = ts.ModuleKind.CommonJS;
+    parsed.options.moduleResolution = ts.ModuleResolutionKind.NodeJs;
+  }
+
+  const host = ts.createCompilerHost(parsed.options, true);
+  return ts.createProgram(parsed.fileNames, parsed.options, host);
+}
