--- packages/dtslint/test/expect/tsconfig.json
@@ -1,1 +1,0 @@
-{}
